+++ database/src/lib/uid.rs
@@ -3,8 +3,8 @@ use std::collections::BTreeSet;
 /// Incrementing uid that that uses the smallest available integer for the next uid or always
 /// increments to the next uid.
 pub struct IntCursor {
-  pos: u128,
-  free: BTreeSet<u128>,
+  pos: u64,
+  free: BTreeSet<u64>,
 }
 
 impl IntCursor {
@@ -15,7 +15,7 @@ impl IntCursor {
     }
   }
 
-  pub fn next(&mut self) -> u128 {
+  pub fn next(&mut self) -> u64 {
     let free = &mut self.free;
 
     if free.len() == 0 {
@@ -26,12 +26,12 @@ impl IntCursor {
   }
 
   /// Returns true if the value existed or false if the value did not exist.
-  pub fn drop(&mut self, pos: u128) -> bool {
+  pub fn drop(&mut self, pos: u64) -> bool {
     self.free.insert(pos);
     pos <= self.pos
   }
 
-  fn increment(&mut self) -> u128 {
+  fn increment(&mut self) -> u64 {
     let pos = self.pos;
     self.pos += 1;
     pos
